Move val_mcm to frontend/test/pytest/test_mid_circuit_measurement.py
