+++ frontend/test/pytest/test_mid_circuit_measurement.py
@@ -12,10 +12,13 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from functools import reduce
+from typing import Iterable, Sequence
+
 import jax.numpy as jnp
+import numpy as np
 import pennylane as qml
 import pytest
-from conftest import validate_measurements
 from pennylane.transforms.dynamic_one_shot import fill_in_value
 
 from catalyst import CompileError, cond, dynamic_one_shot, measure, qjit
@@ -462,6 +465,112 @@ class TestMidCircuitMeasurement:
                 continue
             r1, r0 = qml.math.array(r1).ravel(), qml.math.array(r0).ravel()
             qml.math.allclose(r1, r0)
+
+
+def validate_counts(shots, results1, results2, batch_size=None):
+    """Compares two counts.
+
+    If the results are ``Sequence``s, loop over entries.
+
+    Fails if a key of ``results1`` is not found in ``results2``.
+    Passes if counts are too low, chosen as ``100``.
+    Otherwise, fails if counts differ by more than ``20`` plus 20 percent.
+    """
+    if isinstance(shots, Sequence):
+        assert isinstance(results1, tuple)
+        assert isinstance(results2, tuple)
+        assert len(results1) == len(results2) == len(shots)
+        for s, r1, r2 in zip(shots, results1, results2):
+            validate_counts(s, r1, r2, batch_size=batch_size)
+        return
+
+    if batch_size is not None:
+        assert isinstance(results1, Iterable)
+        assert isinstance(results2, Iterable)
+        assert len(results1) == len(results2) == batch_size
+        for r1, r2 in zip(results1, results2):
+            validate_counts(shots, r1, r2, batch_size=None)
+        return
+
+    for key1, val1 in results1.items():
+        val2 = results2[key1]
+        if abs(val1 + val2) > 100:
+            assert np.allclose(val1, val2, atol=20, rtol=0.2)
+
+
+def validate_samples(shots, results1, results2, batch_size=None):
+    """Compares two samples.
+
+    If the results are ``Sequence``s, loop over entries.
+
+    Fails if the results do not have the same shape, within ``20`` entries plus 20 percent.
+    This is to handle cases when post-selection yields variable shapes.
+    Otherwise, fails if the sums of samples differ by more than ``20`` plus 20 percent.
+    """
+    if isinstance(shots, Sequence):
+        assert isinstance(results1, tuple)
+        assert isinstance(results2, tuple)
+        assert len(results1) == len(results2) == len(shots)
+        for s, r1, r2 in zip(shots, results1, results2):
+            validate_samples(s, r1, r2, batch_size=batch_size)
+        return
+
+    if batch_size is not None:
+        assert isinstance(results1, Iterable)
+        assert isinstance(results2, Iterable)
+        assert len(results1) == len(results2) == batch_size
+        for r1, r2 in zip(results1, results2):
+            validate_samples(shots, r1, r2, batch_size=None)
+        return
+
+    sh1, sh2 = results1.shape[0], results2.shape[0]
+    assert np.allclose(sh1, sh2, atol=20, rtol=0.2)
+    assert results1.ndim == results2.ndim
+    if results2.ndim > 1:
+        assert results1.shape[1] == results2.shape[1]
+    np.allclose(qml.math.sum(results1), qml.math.sum(results2), atol=20, rtol=0.2)
+
+
+def validate_expval(shots, results1, results2, batch_size=None):
+    """Compares two expval, probs or var.
+
+    If the results are ``Sequence``s, validate the average of items.
+
+    If ``shots is None``, validate using ``np.allclose``'s default parameters.
+    Otherwise, fails if the results do not match within ``0.01`` plus 20 percent.
+    """
+    if isinstance(shots, Sequence):
+        assert isinstance(results1, tuple)
+        assert isinstance(results2, tuple)
+        assert len(results1) == len(results2) == len(shots)
+        results1 = reduce(lambda x, y: x + y, results1) / len(results1)
+        results2 = reduce(lambda x, y: x + y, results2) / len(results2)
+        validate_expval(sum(shots), results1, results2, batch_size=batch_size)
+        return
+
+    if shots is None:
+        assert np.allclose(results1, results2)
+        return
+
+    if batch_size is not None:
+        assert len(results1) == len(results2) == batch_size
+        for r1, r2 in zip(results1, results2):
+            validate_expval(shots, r1, r2, batch_size=None)
+
+    assert np.allclose(results1, results2, atol=0.01, rtol=0.2)
+
+
+def validate_measurements(func, shots, results1, results2, batch_size=None):
+    """Calls the correct validation function based on measurement type."""
+    if func is qml.counts:
+        validate_counts(shots, results1, results2, batch_size=batch_size)
+        return
+
+    if func is qml.sample:
+        validate_samples(shots, results1, results2, batch_size=batch_size)
+        return
+
+    validate_expval(shots, results1, results2, batch_size=batch_size)
 
 
 if __name__ == "__main__":
